user-service/requirements.txt
grpcio>=1.65.0
grpcio-tools>=1.65.0
protobuf>=4.25.0
psycopg2-binary>=2.9.0
bcrypt>=4.1.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
requests>=2.31.0